chore(selector): rename preserved default selector
