+++ src/renderer/src/components/Selector/BasicSelector.tsx
